requirements.txt
flask
werkzeug
pillow
numpy
joblib
google-generativeai
python-dotenv
scikit-learn
requests
gunicorn
